examples/taskpilot/requirements.txt
agent-framework>=1.0.0b251001
openai
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
# Guardrails dependencies
nemoguardrails>=0.5.0
requests>=2.31.0
aiohttp>=3.9.0
# Embedded OPA
opa-python>=0.0.9
# Metrics server (for observability)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
# OpenTelemetry for distributed tracing
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp-proto-grpc>=1.21.0
# JSON logging
python-json-logger>=2.0.7
